Add nonEmptyCategories array to Category
The categories are updated on import and a notification is broadcast when it finished updating, to let views know to reload. The categories are also saved in UserDefaults so in case there is no network, and to keep the collection view from being empty on launch.Updated collection view to be based off of the new array, so that it only shows categories with questions.